implement metric_expand
